wrap root route inside an <App/> PureComponent
